Serializer refactor into utils
