+++ pcss-criminal-model/src/main/resources/jaxbConfigs/jaxb-datetime-binding.xjb
@@ -2,7 +2,7 @@
         <jaxb:globalBindings>
             <jaxb:serializable uid="1" />
             <jaxb:javaType name="java.time.Instant" xmlType="xsd:dateTime"
-                   parseMethod="ca.bc.gov.open.pcsscriminalcommon.serializer.InstantSerializer.parseSoap"
-                 printMethod="ca.bc.gov.open.pcsscriminalcommon.serializer.InstantSerializer.print" />
+                   parseMethod="ca.bc.gov.open.pcsscriminalcommon.utils.InstantUtils.parseSoap"
+                 printMethod="ca.bc.gov.open.pcsscriminalcommon.utils.InstantUtils.print" />
         </jaxb:globalBindings>
 </jaxb:bindings>
